+++ src/js/core/form-custom.js
@@ -1,5 +1,5 @@
 import Class from '../mixin/class';
-import {$, $$, includes, isInput, matches, query, selInput, toggleClass} from 'uikit-util';
+import {$, $$, isInput, matches, query, selInput} from 'uikit-util';
 
 export default {
 
@@ -57,36 +57,12 @@ export default {
 
     },
 
-    events: [
+    events: {
 
-        {
-
-            name: 'focusin focusout mouseenter mouseleave',
-
-            delegate: selInput,
-
-            handler({type, current}) {
-                if (current === this.input) {
-                    toggleClass(
-                        this.state,
-                        `uk-${includes(type, 'focus') ? 'focus' : 'hover'}`,
-                        includes(['focusin', 'mouseenter'], type)
-                    );
-                }
-            }
-
-        },
-
-        {
-
-            name: 'change',
-
-            handler() {
-                this.$emit();
-            }
-
+        change() {
+            this.$emit();
         }
 
-    ]
+    }
 
 };