+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fix touch event detection
 - Fix background images are shown too large on retina displays in Img component
+- Fix correctly calculate offsetLeft in Img component
 
 ## 3.0.2 (January 15, 2019)